[WIP] Fix API pause
